+++ packages/@runlightyear/github/src/listeners/onWorkflowRun.ts
@@ -5,6 +5,6 @@ import {
 } from "../webhooks/payloads/asWorkflowRunPayload";
 
 export const onWorkflowRun = createListener<WorkflowRunPayload>({
-  event: "issues",
+  event: "workflow_run",
   payloadCaster: asWorkflowRunPayload,
 });