+++ packages/@runlightyear/gcal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @runlightyear/gcal
 
+## 0.6.0
+
+### Minor Changes
+
+- Updated the typescript target to es2020
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @runlightyear/lightyear@0.30.0
+
 ## 0.5.3
 
 ### Patch Changes