+++ packages/@runlightyear/linear/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @runlightyear/linear
 
+## 0.6.0
+
+### Minor Changes
+
+- Updated the typescript target to es2020
+- Added the findIssueByIdentifier and findWorkflowStateByName helpers
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @runlightyear/lightyear@0.30.0
+
 ## 0.5.10
 
 ### Patch Changes