+++ packages/@runlightyear/smtp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @runlightyear/smtp
 
+## 0.2.0
+
+### Minor Changes
+
+- Updated the typescript target to es2020
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @runlightyear/lightyear@0.30.0
+
 ## 0.1.16
 
 ### Patch Changes